Tweaked some debugging output from old MappingCalculator code. Turns out that the SMALLBANK benchmark should use BIGINTs for all of its customer identifier columns. #117

+++ src/benchmarks/edu/brown/benchmark/smallbank/smallbank-ddl.sql
@@ -1,19 +1,19 @@
 CREATE TABLE ACCOUNTS (
-    custid      INTEGER     NOT NULL,
+    custid      BIGINT      NOT NULL,
     name        VARCHAR(64) NOT NULL,
     CONSTRAINT pk_accounts PRIMARY KEY (custid),
 );
 CREATE INDEX IDX_ACCOUNTS_NAME ON ACCOUNTS (name);    
 
 CREATE TABLE SAVINGS (
-    custid      INTEGER     NOT NULL,
+    custid      BIGINT      NOT NULL,
     bal         FLOAT       NOT NULL,
     CONSTRAINT pk_savings PRIMARY KEY (custid),
     FOREIGN KEY (custid) REFERENCES ACCOUNTS (custid)
 );
 
 CREATE TABLE CHECKING (
-    custid      INTEGER     NOT NULL,
+    custid      BIGINT      NOT NULL,
     bal         FLOAT       NOT NULL,
     CONSTRAINT pk_checking PRIMARY KEY (custid),
     FOREIGN KEY (custid) REFERENCES ACCOUNTS (custid)
